Swipeable MessageItemContainer component - copy DiscussionItemContainer to MessageItemContainer - delete DiscussionItem - update routes - update locales
