+++ src/frontend/web_application/locales/en/main.json
@@ -110,13 +110,6 @@
     "device.verify.descr": "You need to verify this device to allow it to access to your Caliopen account. Please note that this action can have an impact on your Privacy Index according to the device information and settings. You'll find below some informations about the device which tried to connect to your Caliopen account.",
     "device.verify.not-you": "It's not you?",
     "device.verify.title": "Verify you device",
-    "discussion-item.action.archives": "Archives",
-    "discussion-item.action.manage_tags": "Manage tags",
-    "discussion-item.action.enable_tracking": "Enable tracking",
-    "discussion-item.action.delete": "Delete",
-    "discussion-item.action.forward": "Forward",
-    "discussion-item.action.more": "More",
-    "discussion-item.action.reply": "Reply",
     "general.action.cancel": "Cancel",
     "general.action.load_more": "Load more",
     "header.menu.account": "Account",
@@ -143,6 +136,9 @@
     "login.details.action.change": "Change your login",
     "login.details.label": "login",
     "login.details.title": "Login:",
+    "message-item.action.manage_tags": "Manage tags",
+    "message-item.action.delete": "Delete",
+    "message-item.action.reply": "Reply",
     "message-list.action.copy-to": "Copy to",
     "message-list.action.delete": "Delete",
     "message-list.action.reply": "Reply",
