+++ src/frontend/web_application/locales/fr/main.json
@@ -110,13 +110,6 @@
     "device.verify.descr": "Vous devez vérifier cet appareil afin de lui permettre de se connecter à votre compte Caliopen. Merci de noter que cette opération a un impact sur l'Index de Confidentialité selon ses informations et paramétrages. Vous trouverez ci-dessous quelques informations à propos de cet appareil qui a essayé de se connecter à votre compte Caliopen.",
     "device.verify.not-you": "Ce n'est pas le vôtre ?",
     "device.verify.title": "Vérifier votre appareil",
-    "discussion-item.action.archives": "Archives",
-    "discussion-item.action.manage_tags": "Gestion des tags",
-    "discussion-item.action.enable_tracking": "Activer le suivi",
-    "discussion-item.action.delete": "Supprimer",
-    "discussion-item.action.forward": "Faire suivre",
-    "discussion-item.action.more": "Plus",
-    "discussion-item.action.reply": "Répondre",
     "general.action.cancel": "Annuler",
     "general.action.load_more": "Charger plus",
     "header.menu.account": "Compte",
@@ -143,6 +136,9 @@
     "login.details.action.change": "Changer d'identifiant",
     "login.details.label": "Login",
     "login.details.title": "Login:",
+    "message-item.action.manage_tags": "Gestion des tags",
+    "message-item.action.delete": "Supprimer",
+    "message-item.action.reply": "Répondre",
     "message-list.action.copy-to": "Copier à",
     "message-list.action.delete": "Supprimer",
     "message-list.action.reply": "Répondre",
