+++ src/frontend/web_application/locales/template/main.json
@@ -110,13 +110,6 @@
     "device.verify.descr": "device.verify.descr",
     "device.verify.not-you": "device.verify.not-you",
     "device.verify.title": "device.verify.title",
-    "discussion-item.action.archives": "discussion-item.action.archives",
-    "discussion-item.action.manage_tags": "discussion-item.action.manage_tags",
-    "discussion-item.action.enable_tracking": "discussion-item.action.enable_tracking",
-    "discussion-item.action.delete": "discussion-item.action.delete",
-    "discussion-item.action.forward": "discussion-item.action.forward",
-    "discussion-item.action.more": "discussion-item.action.more",
-    "discussion-item.action.reply": "discussion-item.action.reply",
     "general.action.cancel": "general.action.cancel",
     "general.action.load_more": "general.action.load_more",
     "header.menu.account": "header.menu.account",
@@ -143,6 +136,9 @@
     "login.details.action.change": "login.details.action.change",
     "login.details.label": "login.details.label",
     "login.details.title": "login.details.title",
+    "message-item.action.manage_tags": "message-item.action.manage_tags",
+    "message-item.action.delete": "message-item.action.delete",
+    "message-item.action.reply": "message-item.action.reply",
     "message-list.action.copy-to": "message-list.action.copy-to",
     "message-list.action.delete": "message-list.action.delete",
     "message-list.action.reply": "message-list.action.reply",
