+++ src/frontend/web_application/src/routes.jsx
@@ -5,7 +5,6 @@ import { withTranslator } from '@gandi/react-translate';
 import Contact from './scenes/Contact';
 import Auth from './scenes/Auth';
 import Timeline from './scenes/Timeline';
-import DiscussionList from './scenes/DiscussionList';
 import AppRoute from './scenes/AppRoute';
 import NewDraft from './scenes/NewDraft';
 import UserRoute from './scenes/UserRoute';
@@ -42,7 +41,7 @@ export const getRouteConfig = ({ __ }) => [
       {
         path: '/discussions',
         exact: true,
-        component: DiscussionList,
+        component: Timeline,
         app: 'discussion',
       },
       {
