--- src/frontend/web_application/src/scenes/DiscussionList/components/DiscussionItem/index.spec.js
@@ -1,34 +1,0 @@
-// XXX: unit test
-
-// const moduleName = 'caliopenApp';
-//
-describe('component DiscussionList DiscussionItem', () => {
-  it('needs to be implemented', () => {
-    console.warn('TODO');
-  });
-//   let $componentController;
-//
-//   beforeEach(() => {
-//     // TODO: dependency on root app component to remove
-//     angular.mock.module(moduleName);
-//   });
-//
-//   beforeEach(angular.mock.inject((_$componentController_) => {
-//     $componentController = _$componentController_;
-//   }));
-//
-//
-//   it('thread has no unread', () => {
-//     const bindings = { thread: { unread_count: 0 } };
-//     const ctrl = $componentController('threadItem', null, bindings);
-//
-//     expect(ctrl.hasUnread).toEqual(false);
-//   });
-//
-//   it('thread has unread', () => {
-//     const bindings = { thread: { unread_count: 3 } };
-//     const ctrl = $componentController('threadItem', null, bindings);
-//
-//     expect(ctrl.hasUnread).toEqual(true);
-//   });
-});
