--- src/frontend/web_application/src/scenes/DiscussionList/components/ParticipantsIcon/index.spec.js
@@ -1,77 +1,0 @@
-// XXX: unit test
-
-// import discussions from '../index.js';
-//
-describe('component Discussions ContactsIcon', () => {
-  it('needs to be implemented', () => {
-    console.warn('TODO');
-  });
-
-//   let $componentController;
-//
-//   beforeEach(() => {
-//     angular.mock.module(discussions);
-//   });
-//
-//   beforeEach(angular.mock.inject((_$componentController_) => {
-//     $componentController = _$componentController_;
-//   }));
-//
-//   it('thread has 3 contacts', () => {
-//     const bindings = {
-//       thread: {
-//         contacts: [
-//           { title: 'Prof. Farnsworth' },
-//           { title: 'Leela' },
-//           { title: 'Bender' },
-//         ],
-//       },
-//     };
-//     const ctrl = $componentController('contactsIcon', null, bindings);
-//     expect(ctrl.lettersStylesheetClass).toEqual(['m-letter--p', 'm-letter--l', 'm-letter--b']);
-//     expect(ctrl.iconClass).toEqual('m-avatars__letter--3');
-//   });
-//
-//   it('thread has 4 contacts', inject(() => {
-//     const bindings = {
-//       thread: {
-//         contacts: [
-//           { title: 'Prof. Farnsworth' },
-//           { title: 'Leela' },
-//           { title: 'Bender' },
-//           { title: 'Fry' },
-//         ],
-//       },
-//     };
-//     const ctrl = $componentController('contactsIcon', null, bindings);
-//     expect(ctrl.lettersStylesheetClass).toEqual([
-//       'm-letter--p',
-//       'm-letter--l',
-//       'm-letter--b',
-//       'm-letter--f',
-//     ]);
-//     expect(ctrl.iconClass).toEqual('m-avatars__letter--4');
-//   }));
-//
-//   it('thread has many contacts', inject(() => {
-//     const bindings = {
-//       thread: {
-//         contacts: [
-//           { title: 'Prof. Farnsworth' },
-//           { title: 'Leela' },
-//           { title: 'Bender' },
-//           { title: 'Fry' },
-//           { title: 'Dr Zoïdberg' },
-//         ],
-//       },
-//     };
-//     const ctrl = $componentController('contactsIcon', null, bindings);
-//     expect(ctrl.lettersStylesheetClass).toEqual([
-//       'm-letter--p',
-//       'm-letter--l',
-//       'm-letter--b',
-//       'm-letter--plus',
-//     ]);
-//     expect(ctrl.iconClass).toEqual('m-avatars__letter--4');
-//   }));
-});
