--- src/frontend/web_application/src/scenes/DiscussionList/presenter.spec.jsx
@@ -1,36 +1,0 @@
-// XXX: unit test
-
-// // const moduleName = 'caliopenApp';
-//
-describe('component DiscussionList', () => {
-  it('needs to be implemented', () => {
-    console.warn('TODO');
-  });
-//   // let $componentController;
-//
-//   beforeEach(() => {
-//     // TODO: dependency on root app component to remove
-//     // angular.mock.module(moduleName, ($provide, $translateProvider) => {
-//     //   $translateProvider.translations('en', {});
-//     //   $translateProvider.preferredLanguage('en');
-//     //   $provide.decorator('$httpBackend', ($delegate, ApiUrl) => {
-//     //     $delegate.whenRoute('GET', `${ApiUrl}/me`).respond(200, { });
-//     //     $delegate.when('GET', /threads/).respond(200, { threads: [{ foo: 'bar' }] });
-//     //
-//     //     return $delegate;
-//     //   });
-//     // });
-//   });
-//
-//   // beforeEach(angular.mock.inject((_$componentController_) => {
-//   //   $componentController = _$componentController_;
-//   // }));
-//
-//   it('has threads', inject(($rootScope, $httpBackend) => {
-//     const ctrl = $componentController('discussions');
-//     ctrl.$onInit();
-//     $httpBackend.flush();
-//
-//     expect(ctrl.threads).toEqual([{ foo: 'bar' }]);
-//   }));
-});
