+++ src/frontend/web_application/src/scenes/Timeline/components/ManageTags/index.js
@@ -1,11 +1,11 @@
 import { bindActionCreators, compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslator } from '@gandi/react-translate';
-import { updateDiscussion } from '../../../../store/modules/discussion';
+import { updateMessage } from '../../../../store/modules/message';
 import Presenter from './presenter';
 
 const mapDispatchToProps = dispatch => bindActionCreators({
-  onDiscussionChange: updateDiscussion,
+  onMessageChange: updateMessage,
 }, dispatch);
 
 export default compose(
