+++ src/frontend/web_application/src/scenes/Timeline/components/ManageTags/presenter.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import TagsForm from '../../../../components/TagsForm';
+
+const ManageTags = ({ message, onMessageChange }) => {
+  const handleTagsChange = ({ tags }) => {
+    const updateMessage = {
+      ...message,
+      tags,
+    };
+
+    onMessageChange({ message: updateMessage, original: message });
+  };
+
+  return (
+    <TagsForm
+      tags={message.tags}
+      onUpdate={handleTagsChange}
+      onCreate={str => str}
+    />
+  );
+};
+
+ManageTags.propTypes = {
+  message: PropTypes.shape({}).isRequired,
+  onMessageChange: PropTypes.func.isRequired,
+};
+
+export default ManageTags;
