+++ src/frontend/web_application/src/scenes/Timeline/components/MessageItem/presenter.jsx
@@ -2,8 +2,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import Link from '../../../../components/Link';
-// FIXME
-import DiscussionItemContainer from '../../../DiscussionList/components/DiscussionItemContainer';
+import MessageItemContainer from '../MessageItemContainer';
 import AuthorAvatar from '../AuthorAvatar';
 import Icon from '../../../../components/Icon';
 import TextBlock from '../../../../components/TextBlock';
@@ -43,7 +42,7 @@ class MessageItem extends PureComponent {
     const { message, formatDate, __ } = this.props;
 
     return (
-      <DiscussionItemContainer discussion={{}} __={__}>
+      <MessageItemContainer message={{}} __={__}>
         <Link
           to={`/discussions/${message.discussion_id}`}
           className={classnames('s-message-item', { 's-message-item--unread': message.is_unread, 's-message-item--draft': message.is_draft })}
@@ -73,7 +72,7 @@ class MessageItem extends PureComponent {
             </time>
           </div>
         </Link>
-      </DiscussionItemContainer>
+      </MessageItemContainer>
     );
   }
 }
